+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.1.0] - Unreleased
+## [1.1.0] – October 24, 2025
+
+- Add `Cretheus.Decode.integer`
+- Add `Cretheus.Encode.value`
+
+## [1.0.0] – October 24, 2025
 
 - Initial release.